Fixing compiler warnings.

+++ STHorizontalPickerScaleDecimal.m
@@ -92,7 +92,7 @@ const CGFloat kSTHorizontalPickerScaleDecimalHeight = 40.0f;
     CTParagraphStyleSetting paragraphSettings[1];
     
     transform = CGContextGetUserSpaceToDeviceSpaceTransform(context);
-    resolution = sqrtf(fabsf(transform.a * transform.d - transform.b * transform.c)) * 0.5f * (bounds.size.width / imageBounds.size.width + bounds.size.height / imageBounds.size.height);
+    resolution = sqrtf(fabs(transform.a * transform.d - transform.b * transform.c)) * 0.5f * (bounds.size.width / imageBounds.size.width + bounds.size.height / imageBounds.size.height);
     
     CGContextSaveGState(context);
     CGContextTranslateCTM(context, bounds.origin.x, bounds.origin.y);
